+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### 🚀 New features
+- Use [ZLinq](https://github.com/Cysharp/ZLinq) to improve LinQ performances.
 - Added `GenericEqualityComparer`
 
 ### 🧹 Housekeeping